requirements.txt
Django==4.1.2
djangorestframework==3.14.0
drf-yasg==1.21.4
psycopg2-binary==2.9.3
django-debug-toolbar==3.2.4
flake8==4.0.1
requests==2.28.1
pytest==7.1.3
pytest-django==4.5.2
